Add the first aspects of system calls.
This is a very basic ground work for system calls. This is a purely
internal mechanism that does not rely on sysenter/sysexit or interrupts.
However it does lay the ground work to later implement those.

This is about getting a system call resolution mechanism in place, so that
the system call function number is correctly interpreted and sent to the
correct place, and results correctly handed back.